refactor(errors): split error-registry into per-category modules (#1989)
The 842-line error-registry.ts was a single edit hotspot for all 81 error
definitions across 11 categories — merge contention and hard-to-see category
boundaries (audit #1989, finding #7).

Move each category's defineError consts into src/errors/error-registry/<category>.ts
(config, build, runtime, route, module, server, boundary, dev, deploy, agent,
general), each also exporting a <CATEGORY>_REGISTRY slug→definition fragment.
error-registry.ts is now a barrel that re-exports every definition and assembles
ERROR_REGISTRY by spreading the fragments; ErrorSlug, getErrorBySlug,
getErrorsByCategory and getAllSlugs are unchanged.

Pure reorganization — no behavior change. Verified: the exported const set and
the full slug set are byte-identical to before, and the existing
error-registry.test.ts passes unchanged. Refs #1989
